Add refactor of the code
